+++ inst/tinytest/test-gt.R
@@ -1,8 +1,8 @@
 source("helpers.R")
+
 requiet("gt")
 requiet("magrittr")
 using("tinysnapshot")
-# exit_file("gt html snapshot change all the time")
 
 models <- list()
 models[["OLS 1"]] <- lm(hp ~ mpg + wt, mtcars)